fix for tmp directory

+++ tests/ReadFileRepositoryTest.php
@@ -21,7 +21,7 @@ class ReadFileRepositoryTest extends TestCase
     {
         //$this->logs_dir = realpath(dirname(__DIR__, 1)) . DIRECTORY_SEPARATOR . 'logs';
         defined('TMPDIR') or define('TMPDIR', sys_get_temp_dir());
-    
+
         $this->logs_dir = TMPDIR;
         $this->log_name = 'test_logger.log';
         $this->full_path = implode(DIRECTORY_SEPARATOR, [$this->logs_dir, $this->log_name]);
@@ -35,8 +35,10 @@ class ReadFileRepositoryTest extends TestCase
     }
     public function tearDown(): void
     {
-        $deleted = unlink($this->full_path);
-        $this->assertTrue($deleted);
+        if (file_exists($this->full_path)) {
+            $deleted = unlink($this->full_path);
+            $this->assertTrue($deleted);
+        }
         /*
         foreach (glob($this->logs_dir . DIRECTORY_SEPARATOR . '*') as $file) {
             $deleted = unlink($file);
@@ -63,7 +65,7 @@ class ReadFileRepositoryTest extends TestCase
         $content = file_get_contents($this->full_path);
         $this->assertStringContainsString("This is an info", $content);
     }
-    
+
     public function testTriggersDebugInit()
     {
         $debugger = Debugger::set('debug', $this->logger)->initFailuresHandler('debug');
@@ -83,7 +85,7 @@ class ReadFileRepositoryTest extends TestCase
         $content = file_get_contents($this->full_path);
         $this->assertStringContainsString("This is a trigger", $content);
     }
-    
+
     public function testTriggerWarningSupressed()
     {
         $debugger = Debugger::set('error', $this->logger)->initFailuresHandler('error');
